Enable opting-out of form submissions
Similar to [Disabling Turbo Drive on Specific Links][turbo-drive-false],
guard intercepting form subissions with a similar check for
`[data-turbo-drive="false"]` on the submitter, `<form>` element, or an
ancestor.

Similarly, *do not intercept* when the submitter or form element
specifies a method that isn't an HTTP verb (via
[formmethod="dialog"][formmethod] attribute or [method="dialog"][method]
attribute, respectively).

In practice, HTTP submitting `<form>` elements will only ever declare
`method="get"` or `method="post"`. The one [exceptional case is
`method="dialog"`][method-dialog], which we check for as a special case.

Testing
---

Since we're covering behavior for the `<dialog>` element _and_ running
the suite in both Chrome and Firefox, the test suite needs to polyfill
support for the `<dialog>` element.

[turbo-drive-false]: https://turbo.hotwire.dev/handbook/drive#disabling-turbo-drive-on-specific-links
[formmethod]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#attr-formmethod
[method]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form#attr-method
[method-dialog]: https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#form-submission-2