Prevent silent restoration visits when previous visit request is in-flight (#595)
* Prevent silent restoration visits when previous visit request is in-flight.
Silent Visits don't dispatch events. A Visit is deemed to be silent if its location is on the current page. Prior to this commit, this is determined by checking the last rendered location. This approach can result in false positives if a Visit is interrupted and never gets rendered. For example, tapping a link, then immediately navigating back, will be incorrectly considered a same-page visit. Checking the previous location using `window.location` is also not possible on Back/Forward because it is changed _before_ the popstate event, so by the time a new Visit is created, the location has already changed.

This commit accurately compares the last visited location by storing a reference to the `lastVisit` on the Navigator, falling back to the View's `lastRenderedLocation`.

* Obey the linter