Turbo stream source (#415)
* Improve test coverage for streams over SSE

Add a `<form>` element to the `src/tests/fixtures/stream.html` file so
that tests can exercise receiving `<turbo-stream>` elements
asynchronously over an `EventSource` instance that polls for Server-sent
Events.

Extend the `src/tests/server.ts` to account for the `Accept:
text/event-stream` requests that are made by browsers via the `new
EventSource(...)` instance.

* Introduce `<turbo-stream-source>`

Closes https://github.com/hotwired/turbo/issues/413

The `<turbo-stream-source>` element accepts a `[src]` attribute, and
uses that to connect Turbo to poll for streams published on the server
side.

When the element is connected to the document, the stream source is
connected. When the element is disconnected, the stream is disconnected.

When declared with an `ws://` or `wss://` URL, the underlying Stream
Source will be a `WebSocket` instance. Otherwise, the connection is
through an `EventSource`.

Since the document's `<head>` is persistent across navigations, the
`<turbo-stream-source>` is meant to be mounted within the `<body>`
element.

Typical full page navigations driven by Turbo will result in the
`<body>` being discarded and replaced with the resulting document. It's
the server's responsibility to ensure that the element is present on
each page that requires streaming.

* fix lint errors

Co-authored-by: David Heinemeier Hansson <david@hey.com>