Add tests for pausable Frame Rendering (#635)
* Add tests for pausable Frame Rendering

Follow-up to https://github.com/hotwired/turbo/pull/431

Support for the `turbo:before-frame-render` event relies on the same
underlying infrastructure as the `turbo:before-render` event (i.e. the
`Renderer` abstract class). As a result of re-using that abstraction,
listeners for the `turbo:before-frame-render` event can also leverage
the `detail.resume` function in the same way to handle asynchronous
rendering.

The changes made in [hotwired/turbo#431][] excluded test coverage for
that behavior. This commit adds that coverage to guard against
regressions in that behvaior.

[hotwired/turbo#431]: https://github.com/hotwired/turbo/pull/431

* Form Links: Copy `[data-turbo-frame]` from `<a>` to `<form>`

Follow-up to https://github.com/hotwired/turbo/pull/633

The flaky test outlined in [hotwired/turbo#633][] was "flaky", but not
in the sense that was originally suspected. Somehow, it was presenting
as a false negative, failing when we thought it should consistently
pass.

On further inspection, it was _passing when it should consistently
fail_.

This commit addresses the underlying issue by copying any
`[data-turbo-frame]` attributes onto the `<form>` element from the `<a>`
element that is clicked outside of the targeted `<turbo-frame>`.

[hotwired/turbo#633]: https://github.com/hotwired/turbo/pull/633