Fix `assert.equal` in frame tests
The `assert.equal` matcher had been overridden for some tests to allow
string comparisons to act on trimmed versions of the strings.
Unfortunately the overridden matcher was faulty, in that it didn't fail
when the values differed. That is, `assert.equal(1, 2)` was passing
rather than failing.

This change addresses this by:

- Restoring the original `assert.equal`
- Adding a new `assert.equalIgnoringWhitespace` so we can be explicit
  about the cases where we want that behaviour
- Fixing a couple of existing test failures that had been hidden by the
  faulty `assert.equal`