Support View Transition API for navigations (#935)
Adds the ability to use the View Transitions API. It's based around the
MPA View Transitions support which is currently becoming available in
Chrome.

When navigating between pages, we check for the presence of
`view-transition` meta tags in both documents, with a value of
`same-origin`. If those tags are present, and the browser supports the
View Transitions API, we can render the page update within a transition.

When not opted in with the meta tags, or when support is not available,
we fallback to the previous behaviour.

This mimics the behaviour that a supporting browser would have when
performing a full-page navigation between those pages.

We also suppress snapshot caching on pages that specify these meta tags,
since the snapshots would interfere with the animations.

Note that while the API is based on MPA view transitions, the
implementation only requires SPA view transitions, which is already
available in the latest Chrome versions.

Also note that this is based on an API that is very new, and not yet
widely supported. We'll want to keep an eye on how that develops and
update our implementation accordingly.