Add InstantClick behavior (#1101)
* Move doesNotTargetIFrame to util.js

* Move findLinkFromClickTarget to util.js

* Move getLocationForLink to util.js

* Allow request to be intercepted and overriden on turbo:before-fetch-request

* Add instantclick behavior

* Allow customizing the event that triggers prefetching

* Allow customizing the cache time for prefetching

* Rename LinkPrefetchOnMouseoverObserver to LinkPrefetchObserver

Because it is not only triggered on mouseover, but could also be on mousedown, or eventually touchstart.

* Use private methods in LinkPrefetchObserver

* Reorganize methods on LinkPrefetchObserver

* Require a shorter sleep time in the test

Since turbo-prefetch-cache-time is set to 1 millisecond in the html fixture

* Standardize anchor IDs in link_prefetch_observer_tests

anchor_ prefix is used for all anchors in the tests

* Don't try traverse DOM to determine if the target is a link

This is not necessary, since we can just check if the target is an
anchor element with an href attribute.

We were just using findLinkFromClickTarget because it had the selector
we needed, but we can just use the selector directly.

* Keep the closing tag on the same line as the rest of the tag

* Remove unnecessary nesting in tests

* Add missing newline at end of file

* Check for prefetch meta tag before prefetching (on hover event)

* Use FetchRequest to build request for LinkPrefetchObserver

* LinkPrefetchObserver implements the FetchRequest interface, so it can
be used to build a request.
* It also adds this.response to FetchRequest to store the non-awaited
`fetch` response, because we need to FetchRequest#receive() a `fetch`
response, not a FetchRequest.

* Add Turbo Stream header to Accept header when link has data-turbo-stream

* Bring back prefetching links with inner elements

* Add cancelable delay to prefetching links on hover

* Fix clearing cache on every prefetch after b9e82f2

* Add tests for the delay on the meta tag

* Use mouseenter and mouseleave instead of mouseover and mouseout

To avoid having to traverse the DOM to find the link element

* Remove unneeded comment

* Use double quotes instead of single quotes for consistency

* Move link variable declaration inside if statement

Since target is only a link if isLink is true

* Use correct key name for mouseenter event on LinkPrefetchObserver.triggerEvents

On 5078e0b we started using the `mouseenter` event instead of the `mouseover` event to trigger prefetching. However, we forgot to update the key name on the `LinkPrefetchObserver.triggerEvents` object.

* Allow prefetching when visiting page without meta, then visiting one with it

* Allow create and delete posts with comments on the test server

* Clear prefetch cache after form submission

* Add test for nested data-turbo-prefetch=true within data-turbo-prefetch=false

* No longer allow customizing the prefetch trigger event

* No longer allow customizing the prefetch delay

* Add touchstart event to prefetch observer

* Fix flaky tests

This commit fixes the flaky tests by ensuring that each worker has its own database file.

This is done by adding a `worker_id` query parameter to the URLs of the pages that are being tested. This `worker_id` is passed to the database functions, which then use it to determine the name of the database file.

It's necessary because the tests are running in parallel, and the database file is shared between all the workers. This means that if one worker creates a post, the other workers will see that post, and the tests will fail.

* Use double quotes instead of single quotes

* Only cache the link you're currently hovering

Instead of maintaining a cache of all the links that have been hovered
in the last 10 seconds.

This solves issues where the user hovers a link, then performs a non-safe
action and then later clicks the link. In this case, we would be showing
stale content from before the action was performed.

* Remove unused files after ETA template rendered removal

* Remove unused variable

* Clear prefetch cache when the link is no longer hovered

This avoids a flurry of requests when casually scrolling down a page

* Style changes

---------

Co-authored-by: Alberto Fernández-Capel <afcapel@gmail.com>